--- a/haddock-api/src/Haddock/Backends/Xhtml.hs
+++ b/haddock-api/src/Haddock/Backends/Xhtml.hs
@@ -21,6 +21,7 @@ module Haddock.Backends.Xhtml (
 
 import Prelude hiding (div)
 
+import Data.Aeson
 import Data.String
 import GHC.Utils.Error
 import Haddock.Backends.Xhtml.Decl
@@ -440,6 +441,18 @@ instance ToJSON JsonIndexEntry where
         , fromString "link"         .= jieLink
         ]
 
+    toEncoding JsonIndexEntry
+        { jieHtmlFragment
+        , jieName
+        , jieModule
+        , jieLink } =
+      pairs
+        ( fromString "display_html" .= jieHtmlFragment
+        <> fromString "name"         .= jieName
+        <> fromString "module"       .= jieModule
+        <> fromString "link"         .= jieLink
+        )
+
 instance FromJSON JsonIndexEntry where
     parseJSON = withObject "JsonIndexEntry" $ \v ->
       JsonIndexEntry
@@ -485,11 +498,10 @@ ppJsonIndex logger odir maybe_source_url maybe_wiki_url unicode pkg qual_opt ifa
       Builder.hPutBuilder
         h (encodeToBuilder (encodeIndexes (concat installedIndexes)))
   where
-    encodeIndexes :: [JsonIndexEntry] -> Value
+    encodeIndexes :: [JsonIndexEntry] -> [JsonIndexEntry]
     encodeIndexes installedIndexes =
-      toJSON
-        (concatMap fromInterface ifaces
-         ++ installedIndexes)
+        concatMap fromInterface ifaces
+            ++ installedIndexes
 
     fromInterface :: Interface -> [JsonIndexEntry]
     fromInterface iface =