--- a/haddock-api/src/Haddock/Utils/Json.hs
+++ b/haddock-api/src/Haddock/Utils/Json.hs
@@ -11,11 +11,11 @@ module Haddock.Utils.Json
     ( Value(..)
     , Object, object, (.=)
     , encodeToBuilder
-    , ToJSON(toJSON)
+    , ToJSON(..)
 
     , Parser(..)
     , Result(..)
-    , FromJSON(parseJSON)
+    , FromJSON(..)
     , withObject
     , withArray
     , withBool