@ulysses4ever 2022-12-27
RE arson dependency footprint : to fight that, there once was https://github.com/haskell-hvr/microaeson I wonder how it compares performance-wise...