@parsonsmatt 2022-12-27
I'm not sure - #1559 uses `Text` and `Map` and doesn't get close. I suspect the machinery in `toEncoding` and the other `Value -> Builder` is the secret sauce.

I did actually try the `toEncoding . toJSON`, and it was slightly slower, but only ~10% worse than the `toEncoding` variant.

The implementation of `toEncoding` on a `[a]` to pretty simple - 

```haskell
list :: (a -> Encoding) -> [a] -> Encoding
list _  []     = emptyArray_
list to' (x:xs) = openBracket >< to' x >< commas xs >< closeBracket
  where
    commas = foldr (\v vs -> comma >< to' v >< vs) empty
{-# INLINE list #-}
```

`><` is `<>` but pushed into `Encoding`.