@Kleidukos 2024-05-18
Hi, thank you for this PR, but Haddock now lives full-time in the GHC repository! 
Read more at https://discourse.haskell.org/t/haddock-now-lives-in-the-ghc-repository/9576.

Let me know if you feel it is still needed, and I'll migrate it. :)